streamlit_prototyping_special_mission_1/requirements.txt
albumentations==1.1.0
timm==0.6.11
torch==1.13.0
torchvision==0.11.1
numpy==1.23.4
streamlit==1.1.0
Pillow==9.3.0